pyproject.toml
[project]
name = "kernel-tinker-rl"
version = "0.1.0"
description = "Train computer use agents with RL using Kernel browsers and Tinker"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "chz>=0.4.0",
    "cuid2>=2.0.1",
    "kernel>=0.23.0",
    "openai>=2.13.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "tinker-cookbook>=0.1.0",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "ruff>=0.1.0"]
wandb = ["wandb>=0.16.0"]

[project.urls]
Documentation = "https://docs.onkernel.com"
Repository = "https://github.com/onkernel/kernel-tinker-rl"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
